+++ src/overview/src/NIOverviewMemoryCacheController.m
@@ -91,16 +91,16 @@
   if ([self.cache isKindOfClass:[NIImageMemoryCache class]]) {
     NIImageMemoryCache* imageCache = (NIImageMemoryCache *)self.cache;
     summary = [NSString stringWithFormat:
-               @"Number of images: %d\nNumber of pixels: %@/%@\nStress limit: %@",
-               self.cache.count,
+               @"Number of images: %lu\nNumber of pixels: %@/%@\nStress limit: %@",
+               (unsigned long)self.cache.count,
                NIStringFromBytes(imageCache.numberOfPixels),
                NIStringFromBytes(imageCache.maxNumberOfPixels),
                NIStringFromBytes(imageCache.maxNumberOfPixelsUnderStress)];
 
   } else {
     summary = [NSString stringWithFormat:
-               @"Number of objects: %d",
-               self.cache.count];
+               @"Number of objects: %lu",
+               (unsigned long)self.cache.count];
   }
   [contents addObject:[NITableViewModelFooter footerWithTitle:summary]];
