+++ src/photos/src/NIToolbarPhotoViewController.m
@@ -443,9 +443,9 @@
 
 ///////////////////////////////////////////////////////////////////////////////////////////////////
 - (void)setChromeTitle {
-  self.title = [NSString stringWithFormat:@"%d of %d",
-                (self.photoAlbumView.centerPageIndex + 1),
-                self.photoAlbumView.numberOfPages];
+  self.title = [NSString stringWithFormat:@"%ld of %ld",
+                (long)(self.photoAlbumView.centerPageIndex + 1),
+                (long)self.photoAlbumView.numberOfPages];
 }
